Moved CoffeeEventRepository interface to the same file RealmCoffeeEventRepository resides in.
