+++ app/src/main/java/com/codemate/koffeemate/data/network/models/Profile.kt
@@ -6,7 +6,6 @@ open class Profile(
         open var first_name: String = "",
         open var last_name: String = "",
         open var real_name: String = "",
-
         open var image_72: String? = null,
         open var image_192: String? = null,
         open var image_512: String? = null
