+++ app/src/main/java/com/codemate/koffeemate/data/network/models/User.kt
@@ -12,7 +12,6 @@ open class User(
         open var id: String = "",
         open var name: String = "",
         open var profile: Profile = Profile(),
-
         open var real_name: String? = null,
         open var is_bot: Boolean = false,
         open var deleted: Boolean = false
